feat: Introduce a new permission system, adding permission checks to financial operations and team management use cases.

+++ app/(app)/income/_use_case/delete-income.use-case.ts
@@ -1,10 +1,22 @@
-import type { IIncomeRepository } from '@/domain/interfaces/income.repository.interface'
-import type { DeleteIncomeDTO } from '@/domain/dto/income.types.d.ts'
+import type { IIncomeRepository } from "@/domain/interfaces/income.repository.interface";
+import type { DeleteIncomeDTO } from "@/domain/dto/income.types.d.ts";
+
+import type { ITeamRepository } from "@/domain/interfaces/team.repository.interface";
 
 export class DeleteIncomeUseCase {
-  constructor(private incomeRepository: IIncomeRepository) {}
+  constructor(
+    private incomeRepository: IIncomeRepository,
+    private teamRepository: ITeamRepository
+  ) {}
 
   async execute(dto: DeleteIncomeDTO): Promise<void> {
-    await this.incomeRepository.delete(dto.incomeId, dto.teamId)
+    const hasPermission = await this.teamRepository.verifyPermission(
+      dto.userId,
+      dto.teamId,
+      "MANAGE_BUDGET"
+    );
+    if (!hasPermission) throw new Error("Permissão negada.");
+
+    await this.incomeRepository.delete(dto.incomeId, dto.teamId);
   }
 }
